fix problem with unsigned conversion

+++ llimits.h
@@ -262,7 +262,7 @@ union luai_Cast { double l_d; LUA_INT32 l_p[2]; };
 #define lua_number2unsigned(i,n)  \
 	((i)=(lua_Unsigned)((n) - floor((n)/SUPUNSIGNED)*SUPUNSIGNED))
 #else
-#define lua_number2unsigned(i,n)	((i)=(lua_Unsigned)(n))
+#define lua_number2unsigned(i,n)	((i)=(lua_Unsigned)(n.toInt()))
 #endif
 #endif				/* } */
